% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/daycount.R
\name{edate}
\alias{edate}
\title{Shift date by a number of months}
\usage{
edate(from, months = 1)
}
\arguments{
\item{from}{starting date - a character string or any object that can be
converted into date using \code{\link{as.Date}}.}

\item{months}{Number of months (can be negative)}
}
\description{
Convenience function for finding the same date in different months. Used for example
to find coupon dates of bonds given the maturity date. See \code{\link{coupons}}
}

