% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcf.R
\name{equiv.rate}
\alias{equiv.rate}
\title{Equivalent Rates under different Compounding Conventions}
\usage{
equiv.rate(rate, from.freq = 1, to.freq = 1)
}
\arguments{
\item{rate}{The interest rate in decimal (0.10 or 10e-2 for 10\%)}

\item{from.freq}{Frequency of compounding of the given interest
rate: 1 for annual, 2 for semi-annual, 12 for monthly, Inf for
continuous compounding.}

\item{to.freq}{Frequency of compounding to which the given
interest rate is to be converted: 1 for annual, 2 for
semi-annual, 12 for monthly, \code{Inf} for continuous
compounding.}
}
\description{
Converts an interest rate from one compounding convention to
another (for example from semi-annual to monthly compounding or
from annual to continuous compounding)
}

